+++ CHANGELOG.md
@@ -3,7 +3,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 2.0.0-alpha - 2020-11-08
+## 2.0.0.1 - 06-12-2020
 
 ### Added
-- SmsFoundation - Send Sms for customers
+- New Feature `getMessage()` has been added.
+
+### Fixed
+- Partially made compatible with PHP8.
+
+## 2.0.0-alpha - 08-11-2020
+
+### Added
+- SmsFoundation - Send Sms for customers.
+
+### Change
+- setMessage change sendMessage on SmsFoundation.